monkeypatch the activerecord connection pool so that it does not reconnect after forking automatically
nitra manages this itself - it reconnects when it spawns a new worker (which we do need to do), and not when forking again to run a new spec file (which is not optimal because then the activerecord fixture cache resets).

older versions of activerecord didn't use this per-pid pool caching mechanism, and so didn't automatically reconnect.  this monkeypatch effectively restores the old behavior.