Add limited support for splitting Cucumber scenario outlines
The main obstacle here is that ExampleRow objects don't have locations
in their own right – they are subordinate to the scenario outline. So,
use those outlines and dedupe the resultant `file:line` patterns.

This is fairly tightly-coupled to the internal Cucumber design, to the
point that this code effectively partitions on *class*, save that the
`respond_to?`-centric approach does not immediately break under a
modification of class layout (as with Cucumber v2).