Ensure that the test suite passes under rake again
At the same time, we should probably explictly depend on rake (in dev)
given that we're maintaining a rakefile.

This came down to some discrepancies over what was in ARGV between the
run modes: when running an independent file (via e.g. `ruby`), the array
is (understandably) empty, and this caused `configuration.frameworks` to
be called (and consequentially `#slave_mode`).

In contrast, when run via `rake test`, ARGV is the original argument
array to rake (i.e. it is ['test']); this leads to a significant change
in what methods are invoked. Adapting the test in `command_line.rb` to
use `argv` (the initialised attribute) corrects the vast majority of
these tests.

The other case, for file lists, simply does not invoke the relevant
`#framework` and `#slave_mode` methods, as `argv` is not empty during
this test run (whereas ARGV _is_). This is a mocked object, so calls to
these would fail without the expectations, so we can just drop the
offending lines.